// src/app/(protected)/master-data/targets/types.d.ts
export type TargetItem = {
	id: number
	countryImageSrc: string
	comapanyLogoSrc: string
	ticker: string
	companyName: string
	industry: string
	sector: string
	marketCap: string
	revenueGrowth: number
	forwardPe: number
	forwardEvEbitda: number
	priceBook: number
	ltmEvSales: number
	ltmEbidtaOperatingMargin: number
	fiveYearTsr: number
	threeYearTsr: number
	twoYearTsr: number
	oneYearTsr: number
	sixMonthsTsr: number
}